// models/enums.ts
export enum authProviderEnum {
  credentials = 'credentials',
};

export enum BlobTypeEnum {
  image = "1",
  file = "2"
}

export enum productPagesEnum {
  futures = "futures",
  competitors = "competitors",
  factors = "factors",
  ideaFactors = "idea-factors",
}

export enum videoPropNamesEnum {
  introductoryVideo = "introductoryVideo",
  goalsVideo = "goalsVideo",
  staffOnDemand = "staffOnDemand",
  communityAndCrowd = "communityAndCrowd",
  algorithms = "algorithms",
  leveragedAssets = "leveragedAssets",
  Engagement = "Engagement",
  interface = "interface",
  dashboard = "dashboard",
  experimentation = "experimentation",
  socialPlatforms = "socialPlatforms",
  ecoSystems = "ecoSystems",
  autonomy = "autonomy",
  infoIsPower = "infoIsPower",
  OTCR = "OTCR",
  valueDestruction = "valueDestruction",
  customerJourney = "customerJourney",
  digitalPlatforms = "digitalPlatforms",
  buildingCapacity = "buildingCapacity",
  products = "products",
  marketPotential = "marketPotential",
  redOcean = "redOcean",
  voiceOfCustomers = "voiceOfCustomers",
  blueOcean = "blueOcean",
  nonCustomers = "nonCustomers",
  stepUpStepDownModel = "stepUpStepDownModel",
  roadMap = "roadMap",
}

export enum stepNamesEnum {
  visualizeSuccess = "Visualize Success",
  pioneerMigratorSettler = "Pioneer Migrator Settler",
  marketPotential = "Market Potential",
  redOceanCanvas = "Red Ocean Canvas",
  disruption = "Disruption",
  voiceOfCustomers = "Voice of Customers",
  blueOceanCanvas = "Blue ocean",
  nonCustomers = "Non customers",
  stepUpStepDownModel = "Step Up Step Down Model",
  roadMap = "Roadmap",
}

export enum takeawayTypeEnums {
  scale = "on-scale",
  ideas = "on-ideas",
}

export enum navbarDirectionsEnum {
  vertical = "vertical",
  horizontal = "horizontal",
}

export enum tooltipDirectionsEnum {
  left = "left",
  top = "top",
  right = "right",
  bottom = "bottom",
}